requirements.txt
actionlib==1.13.2
addict==2.4.0
angles==1.9.13
anyio==3.5.0
appdirs==1.4.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
autobahn==17.10.1
Automat==0.8.0
Babel==2.9.1
backcall==0.2.0
base-local-planner==1.17.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
bondpy==1.8.6
bosdyn-api==3.1.0
bosdyn-choreography-client==3.1.0
bosdyn-choreography-protos==3.1.0
bosdyn-client==3.1.0
bosdyn-core==3.1.0
bosdyn-mission==3.1.0
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.4.24
catkin-pkg-modules==0.4.24
cbor==1.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
colorama==0.4.3
constantly==15.1.0
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.0
cycler==0.11.0
Cython==0.29.14
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distlib==0.3.0
distro==1.7.0
docutils==0.18.1
dynamic-reconfigure==1.7.2
empy==3.3.4
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.0.12
fonttools==4.31.2
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
graphviz==0.19.1
grpcio==1.44.0
httplib2==0.14.0
hyperlink==19.0.0
idna==3.3
image-geometry==1.16.0
importlib-metadata==1.5.0
importlib-resources==5.6.0
incremental==16.10.1
interactive-markers==1.12.0
ipykernel==6.12.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.2.1
jupyter-core==4.9.2
jupyter-packaging==0.12.0
jupyter-server==1.16.0
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.12.0
jupyterlab-widgets==1.1.0
keyboard==0.13.5
keyring==18.0.1
kiwisolver==1.4.2
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
message-filters==1.15.14
mistune==0.8.4
more-itertools==4.2.0
mpi4py==3.0.3
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
netifaces==0.10.4
nose==1.3.7
notebook==6.4.10
notebook-shim==0.1.0
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
open3d==0.13.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.14.0
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.21
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.2
pygame==2.1.2
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
PyQt5-Qt5==5.15.2
pyRFC3339==1.1
pyrsistent==0.18.1
PySide2==5.15.2.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-gnupg==0.4.5
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.27.1
requests-unixsocket==0.2.0
resource-retriever==1.12.7
roman==2.0.0
rosapi==0.11.13
rosbag==1.15.14
rosboost-cfg==1.15.8
rosbridge-library==0.11.13
rosbridge-server==0.11.13
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.21.0
rosdep-modules==0.21.0
rosdistro==0.8.3
rosdistro-modules==0.8.3
rosgraph==1.15.14
rosinstall==0.7.8
roslaunch==1.15.14
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.14
rosmake==1.15.8
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospkg==1.4.0
rospkg-modules==1.4.0
rospy==1.15.14
rosservice==1.15.14
rostest==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.2
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rtabmap-ros==0.20.18
rviz==1.14.14
scikit-learn==1.0.2
scipy==1.8.0
SecretStorage==2.3.1
Send2Trash==1.8.0
sensor-msgs==1.13.1
service-identity==18.1.0
shiboken2==5.15.2.1
simplejson==3.16.0
sip==4.19.21
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
sniffio==1.2.0
soupsieve==2.3.2
ssh-import-id==5.10
stack-data==0.2.0
systemd-python==234
terminado==0.13.3
testpath==0.6.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
threadpoolctl==3.1.0
tomlkit==0.10.1
topic-tools==1.15.14
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
Twisted==18.9.0
txaio==2.10.0
u-msgpack-python==2.1
urllib3==1.26.9
vcstools==0.1.42
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
widgetsnbextension==3.6.0
wrapt==1.14.0
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.13
zipp==3.8.0
zope.interface==4.7.1
